gtm-prod1/api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
python-telegram-bot==20.6
gunicorn==21.2.0
